refactor(rain): migrate remaining server actions to client-side service
Inline the 6 pre-existing server actions in app/actions/rain.ts
(getRainCardOverview, getRainSessionKeyAddress, submitRainWithdraw-
SessionApproval, prepareRainWithdrawal, submitRainWithdrawal,
stampRainWithdrawal) into services/rain.ts using the existing
rainRequest helper + JWT cookie pattern. Drop app/actions/rain.ts.

All card calls now go through a single client-side rainApi object —
no `'use server'` indirection left for Rain. useGrantSessionKey
switches from the {data, error} return shape to try/catch.
